api/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
blinker==1.6.2
boto3==1.26.129
botocore==1.29.129
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.7
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
factory-boy==3.2.1
Faker==18.10.1
faker-commerce==1.0.3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.3.2
jwt==1.3.1
Markdown==3.4.3
marshmallow==3.19.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
yarg==0.1.9
